video-analysis-backend/requirements.txt
fastapi
uvicorn
celery
redis
python-multipart
pydantic
aiofiles
whisper
opencv-python
numpy
torch
torchvision 
torchaudio
transformers
faster_whisper
pillow
moviepy==1.0.3
deepface
tensorflow
tf-keras
tqdm
pandas
scikit-learn